pyproject.toml
[project]
name = "senile-rod"
version = "0.1.0"
description = ""
authors = [
    {name = "Dominik Rathmer",email = "dominik@rathmer.me"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "platformdirs (>=4.3.8,<5.0.0)",
    "tortoise-orm[asyncpg] (>=0.25.1,<0.26.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
mypy = "^1.16.1"
flake8 = "^7.2.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
senile-rod = "senile_rod.bot:main"